requirements.txt
agno==1.5.6
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
cohere==5.15.0
coloredlogs==15.0.1
deprecation==2.1.0
distro==1.9.0
docstring_parser==0.16
duckduckgo_search==8.0.2
fastapi==0.115.12
fastavro==1.11.1
filelock==3.18.0
flatbuffers==25.2.10
fsspec==2025.5.1
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.40.2
google-genai==1.18.0
h11==0.16.0
hf-xet==1.1.2
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.32.3
humanfriendly==10.0
idna==3.10
Jinja2==3.1.6
jiter==0.10.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
lancedb==0.22.1
lxml==5.4.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
narwhals==1.41.0
numpy==2.2.6
ollama==0.5.1
onnxruntime==1.22.0
openai==1.83.0
opencv-python==4.11.0.86
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pillow==11.2.1
primp==0.15.0
protobuf==6.31.1
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyclipper==1.3.0.post6
pydantic==2.11.5
pydantic-settings==2.9.1
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.1
pylance==0.28.0
pypdf==5.6.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
rapidocr-onnxruntime==1.2.3
referencing==0.36.2
requests==2.32.3
rich==14.0.0
rpds-py==0.25.1
rsa==4.9.1
shapely==2.1.1
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
streamlit==1.45.1
streamlit-chat==0.1.1
sympy==1.14.0
tantivy==0.24.0
tenacity==9.1.2
tokenizers==0.21.1
toml==0.10.2
tomli==2.2.1
tornado==6.5.1
tqdm==4.67.1
typer==0.16.0
types-requests==2.32.0.20250602
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.2
websockets==15.0.1
